chore: remove some inline modules
follow [P-API](https://epage.github.io/dev/rust-style/#p-api)
